requirements.txt
sentence-transformers
faiss-cpu
numpy
python-dotenv
openai
tqdm
flask
youtube-transcript-api
